backend/app/services/gemini/client.py
import os
import google.generativeai as genai
from typing import Dict, List, Optional, Any
import json
from dotenv import load_dotenv

load_dotenv()

class GeminiClient:
    def __init__(self, api_key: Optional[str] = None):
        self.api_key = api_key or os.getenv("GEMINI_API_KEY")
        if not self.api_key:
            raise ValueError("GEMINI_API_KEY not found in environment variables")
        
        genai.configure(api_key=self.api_key)
        self.flash_model = genai.GenerativeModel('gemini-2.0-flash-exp') # Using latest flash
        self.pro_model = genai.GenerativeModel('gemini-1.5-pro')
    
    async def analyze_video(
        self,
        video_file_path: Optional[str] = None,
        video_url: Optional[str] = None,
        prompt: str = "",
        use_pro: bool = False
    ) -> Dict[str, Any]:
        model = self.pro_model if use_pro else self.flash_model
        
        content = []
        
        if video_file_path:
            # For local files, we'd upload them to Gemini's File API first
            # Simplified for now
            video_file = genai.upload_file(path=video_file_path)
            content.append(video_file)
        elif video_url:
            # Gemini can't directly take YouTube URLs in the API without special handling
            # In a real app, we'd download the video or use a specific integration
            # For this hackathon demo, we'll assume we have a way to provide the video
            pass
            
        content.append(prompt)
        
        # In a real scenario, we wait for video processing
        # response = model.generate_content(content)
        
        # Return mock for now if no video provided, otherwise call API
        # if not video_file_path and not video_url:
        return {"status": "success", "message": "Ready for analysis"}
