backend/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
pydantic-settings
sqlalchemy
python-multipart
requests
python-dotenv
reportlab
python-docx
matplotlib
pandas
jinja2
aiohttp
